requirements.txt
jupyterlab
ipykernel
numpy
pandas
scikit-learn
matplotlib
python-dotenv
transformers
sentence-transformers
chromadb
faiss-cpu
openai
tiktoken
langchain
pydantic
pytest
chromadb
openai